requirements.txt
# Core dependencies
requests>=2.31.0
urllib3>=1.26.0

# Binance API
python-binance>=1.0.19

# Telegram Bot API  
python-telegram-bot>=20.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0

# Scheduling and timing
schedule>=1.2.0

# Logging
colorlog>=6.7.0
coloredlogs>=15.0

# Async support
aiohttp>=3.8.0

# Date/time handling
pytz>=2023.3

# Optional: For enhanced features
# plotly>=5.15.0  # For charts
# matplotlib>=3.7.0  # For plotting  
# ta-lib>=0.4.26  # Technical analysis
# cryptography>=41.0.0  # Crypto utilities